+++ src/index.ts
@@ -2,6 +2,8 @@
 export { RabbitMesh } from "./core/rabbitmesh.js";
 export { DLQHandler } from "./core/dlq-handler.js";
 export { DelayHandler } from "./core/delay-handler.js";
+export { ExchangeManager } from "./core/exchange-manager.js";
+export { BindingManager } from "./core/binding-manager.js";
 
 // Interfaces
 export type { RabbitMeshConfig } from "./interfaces/config.interface.js";
@@ -10,6 +12,7 @@ export type { SubscribeOptions, DLQOptions } from "./interfaces/subscribe.interf
 
 // Types
 export type { LogLevel, ResolvedConfig } from "./types/index.js";
+export type { ExchangeType } from "./core/exchange-manager.js";
 
 // Errors
 export {
@@ -19,4 +22,6 @@ export {
   SerializationError,
   RetryError,
   DelayError,
+  ExchangeError,
+  ValidationError,
 } from "./utils/errors.js";